+++ operator/internal/manifests/indexgateway.go
@@ -74,7 +74,6 @@ func BuildIndexGateway(opts Options) ([]client.Object, error) {
 func NewIndexGatewayStatefulSet(opts Options) *appsv1.StatefulSet {
 	l := ComponentLabels(LabelIndexGatewayComponent, opts.Name)
 	a := commonAnnotations(opts.ConfigSHA1, opts.CertRotationRequiredAt)
-
 	podSpec := corev1.PodSpec{
 		Affinity: configureAffinity(LabelIndexGatewayComponent, opts.Name, opts.Gates.DefaultNodeAffinity, opts.Stack.Template.IndexGateway),
 		Volumes: []corev1.Volume{