+++ pkg/storage/stores/indexshipper/downloads/table_manager.go
@@ -30,6 +30,7 @@ const (
 type Limits interface {
 	AllByUserID() map[string]*validation.Limits
 	DefaultLimits() *validation.Limits
+	VolumeMaxSeries(userID string) int
 }
 
 // TenantFilter is invoked by an IndexGateway instance and answers which
@@ -293,6 +294,7 @@ func (tm *tableManager) ensureQueryReadiness(ctx context.Context) error {
 		return nil
 	}
 
+	tm.indexStorageClient.RefreshIndexTableNamesCache(ctx)
 	tables, err := tm.indexStorageClient.ListTables(ctx)
 	if err != nil {
 		return err