+++ pkg/storage/stores/indexshipper/downloads/table_manager_test.go
@@ -405,6 +405,7 @@ func TestTableManager_loadTables(t *testing.T) {
 type mockLimits struct {
 	queryReadyIndexNumDaysDefault int
 	queryReadyIndexNumDaysByUser  map[string]int
+	volumeMaxSeries               int
 }
 
 func (m *mockLimits) AllByUserID() map[string]*validation.Limits {
@@ -422,6 +423,10 @@ func (m *mockLimits) DefaultLimits() *validation.Limits {
 	return &validation.Limits{
 		QueryReadyIndexNumDays: m.queryReadyIndexNumDaysDefault,
 	}
+}
+
+func (m *mockLimits) VolumeMaxSeries(_ string) int {
+	return m.volumeMaxSeries
 }
 
 type mockTable struct {
@@ -464,6 +469,8 @@ func (m *mockIndexStorageClient) ListTables(_ context.Context) ([]string, error)
 func (m *mockIndexStorageClient) ListFiles(_ context.Context, tableName string, _ bool) ([]storage.IndexFile, []string, error) {
 	return []storage.IndexFile{}, m.userIndexesInTables[tableName], nil
 }
+
+func (m *mockIndexStorageClient) RefreshIndexTableNamesCache(_ context.Context) {}
 
 func buildTableNumber(idx int) int64 {
 	return getActiveTableNumber() - int64(idx)