+++ pkg/storage/stores/series/series_index_gateway_store.go
@@ -130,12 +130,13 @@ func (c *IndexGatewayClientStore) Stats(ctx context.Context, userID string, from
 	return resp, nil
 }
 
-func (c *IndexGatewayClientStore) SeriesVolume(ctx context.Context, userID string, from, through model.Time, limit int32, matchers ...*labels.Matcher) (*logproto.VolumeResponse, error) {
+func (c *IndexGatewayClientStore) SeriesVolume(ctx context.Context, userID string, from, through model.Time, limit int32, targetLabels []string, matchers ...*labels.Matcher) (*logproto.VolumeResponse, error) {
 	resp, err := c.client.GetSeriesVolume(ctx, &logproto.VolumeRequest{
-		From:     from,
-		Through:  through,
-		Matchers: (&syntax.MatchersExpr{Mts: matchers}).String(),
-		Limit:    limit,
+		From:         from,
+		Through:      through,
+		Matchers:     (&syntax.MatchersExpr{Mts: matchers}).String(),
+		Limit:        limit,
+		TargetLabels: targetLabels,
 	})
 	if err != nil {
 		if isUnimplementedCallError(err) && c.fallbackStore != nil {
@@ -143,7 +144,7 @@ func (c *IndexGatewayClientStore) SeriesVolume(ctx context.Context, userID strin
 			// Note: this is likely a noop anyway since only
 			// tsdb+ enables this and the prior index returns an
 			// empty response.
-			return c.fallbackStore.SeriesVolume(ctx, userID, from, through, limit, matchers...)
+			return c.fallbackStore.SeriesVolume(ctx, userID, from, through, limit, targetLabels, matchers...)
 		}
 		return nil, err
 	}