+++ pkg/util/marshal/legacy/marshal_test.go
@@ -7,6 +7,7 @@ import (
 	"time"
 
 	json "github.com/json-iterator/go"
+	"github.com/prometheus/prometheus/model/labels"
 	"github.com/stretchr/testify/require"
 
 	loghttp "github.com/grafana/loki/pkg/loghttp/legacy"
@@ -27,6 +28,14 @@ var queryTests = []struct {
 						Timestamp: mustParse(time.RFC3339Nano, "2019-09-13T18:32:22.380001319Z"),
 						Line:      "super line",
 					},
+					{
+						Timestamp: mustParse(time.RFC3339Nano, "2019-09-13T18:32:23.380001319Z"),
+						Line:      "super line with labels",
+						NonIndexedLabels: labels.Labels{
+							{Name: "foo", Value: "a"},
+							{Name: "bar", Value: "b"},
+						},
+					},
 				},
 				Labels: `{test="test"}`,
 			},
@@ -39,6 +48,14 @@ var queryTests = []struct {
 						{
 							"ts": "2019-09-13T18:32:22.380001319Z",
 							"line": "super line"
+						},
+						{
+							"ts": "2019-09-13T18:32:23.380001319Z",
+							"line": "super line with labels",
+							"nonIndexedLabels": {
+								"foo": "a",
+								"bar": "b"
+							}
 						}
 					]
 				}
@@ -164,6 +181,14 @@ var tailTests = []struct {
 							Timestamp: mustParse(time.RFC3339Nano, "2019-09-13T18:32:22.380001319Z"),
 							Line:      "super line",
 						},
+						{
+							Timestamp: mustParse(time.RFC3339Nano, "2019-09-13T18:32:23.380001319Z"),
+							Line:      "super line with labels",
+							NonIndexedLabels: labels.Labels{
+								{Name: "foo", Value: "a"},
+								{Name: "bar", Value: "b"},
+							},
+						},
 					},
 					Labels: "{test=\"test\"}",
 				},
@@ -183,6 +208,14 @@ var tailTests = []struct {
 						{
 							"ts": "2019-09-13T18:32:22.380001319Z",
 							"line": "super line"
+						},
+						{
+							"ts": "2019-09-13T18:32:23.380001319Z",
+							"line": "super line with labels",
+							"nonIndexedLabels": {
+								"foo": "a",
+								"bar": "b"
+							}						
 						}
 					]
 				}